requirements.txt
# Core dependencies
typer>=0.9.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
aiohttp>=3.9.0
jinja2>=3.1.0
genanki>=0.13.0
tiktoken>=0.5.0
markdown>=3.5.0
python-frontmatter>=1.0.0
chardet>=5.0.0
python-dotenv>=1.0.0
loguru>=0.7.0
tqdm>=4.66.0
pyyaml>=6.0.0

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# Optional LLM provider dependencies
openai>=1.0.0
anthropic>=0.18.0
